split datapipe.core_steps
