+++ examples/one_to_many_pipeline/app.py
@@ -3,8 +3,9 @@ from sqlalchemy.sql.schema import Column
 from sqlalchemy.sql.sqltypes import JSON, Boolean, Integer, String
 
 from datapipe.compute import Catalog, DatapipeApp, Pipeline, Table
-from datapipe.core_steps import BatchGenerate, BatchTransform
 from datapipe.datatable import DataStore
+from datapipe.step.batch_generate import BatchGenerate
+from datapipe.step.batch_transform import BatchTransform
 from datapipe.store.database import DBConn, TableStoreDB
 
 dbconn = DBConn("sqlite+pysqlite3:///db.sqlite")
